+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Release 5.2.3:
  - Be more lenient in parsing OpenId authentication requests
+ - OpenID4VP: Use correct format of algorithms in metadata for `vp_formats.vc+sd-jwt`
 
 Release 5.2.2:
  - Remote qualified electronic signatures: